+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Release 0.13.3
+- Fixed a bad internal assert.
+
 ## Release 0.13.2
 - Added `Endpoint::from_listener()` that allows to send datagrams from listeners in non connection-oriented protocols (see the API docs).
 - Removed `'static` restriction in the `NodeListener::for_each()` callback.